Remove redundant coordinator
